Introduce a mechanism to represent model input data as int32 values - Values originally represented as 0 or 1 are encoded as bit values - Values originally represented as real numbers are encoded as integer values Input data represented as int32 values is converted to float32 after being transferred to the GPU. These changes improve the throughput when passing input data to the GPU.

+++ src/deepgo/native/cpp/Board.h
@@ -162,7 +162,7 @@ class Board {
    * @param rule Rule for determining the winner
    * @param superko True to apply the superko rule
    */
-  void getInputs(float* inputs, int32_t color, float komi, int32_t rule, bool superko);
+  void getInputs(int32_t* inputs, int32_t color, float komi, int32_t rule, bool superko);
 
   /**
    * Gets the state of the board.
