+++ src/deepgo/native/cpp/Executor.h
@@ -38,7 +38,7 @@ class Executor {
    * @param outputs Output data
    * @param size Number of input/output data
    */
-  void execute(float* inputs, float* outputs, int32_t size);
+  void execute(int32_t* inputs, float* outputs, int32_t size);
 
   /**
    * Returns the number of pending computation tasks.
